POSTPONE - use comma to store xt of non immediate words after COMPILE and use COMPILE, only for immediate words.
This doesn't make any difference in the compiled code in this naive
implementation, but is more clean semantically.  Consider an
implementation that could do inlining.  COMPILE, may inline the code
of the postponed immediate word here, at the call site.  OTOH we need
to store the xt of the non-immediate word and then COMPILE in its turn
may inline the code later.