Do not abuse host's immediate flag to track target's immediate flag.
Steal a bit from the version field to track target's immediate flag.
Note that when we want to redefine an immediate word, this code will
NOT set clear the immediate flag for the new word.  This is
technically wrong, but the redefinition better be immediate anyway (as
it needs an immediate counterpart in the transpiler to handle it).
Anyway, this will do for now to handle smart IF

XXX: This deserves an explanatory comment in the sources.